// react-frontend/src/api/models/user-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { GrantedAuthority } from './granted-authority';

/**
 * 
 * @export
 * @interface UserEntity
 */
export interface UserEntity {
    /**
     * 
     * @type {number}
     * @memberof UserEntity
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof UserEntity
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserEntity
     */
    'lastName'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserEntity
     */
    'email'?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof UserEntity
     */
    'roles'?: Set<RolesEnum>;
    /**
     * 
     * @type {boolean}
     * @memberof UserEntity
     */
    'accountNonExpired'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UserEntity
     */
    'accountNonLocked'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UserEntity
     */
    'credentialsNonExpired'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UserEntity
     */
    'enabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UserEntity
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserEntity
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserEntity
     */
    'username'?: string;
    /**
     * 
     * @type {Array<GrantedAuthority>}
     * @memberof UserEntity
     */
    'authorities'?: Array<GrantedAuthority>;
}

export const RolesEnum = {
    USER: 'USER',
    ADMIN: 'ADMIN',
    FINANCIAL_ADMIN: 'FINANCIAL_ADMIN',
    MENU_ADMIN: 'MENU_ADMIN',
    USER_ADMIN: 'USER_ADMIN',
    AUDIT_ADMIN: 'AUDIT_ADMIN',
    REPORTING_ADMIN: 'REPORTING_ADMIN',
    FEEDBACK_ADMIN: 'FEEDBACK_ADMIN',
    SYSTEM_ADMIN: 'SYSTEM_ADMIN',
    ORDER_ADMIN: 'ORDER_ADMIN',
    DELIVERY_ADMIN: 'DELIVERY_ADMIN'
} as const;

export type RolesEnum = typeof RolesEnum[keyof typeof RolesEnum];