// react-frontend/src/api/models/admin-update-user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AdminUpdateUserRequest
 */
export interface AdminUpdateUserRequest {
    /**
     * 
     * @type {string}
     * @memberof AdminUpdateUserRequest
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminUpdateUserRequest
     */
    'lastName'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminUpdateUserRequest
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminUpdateUserRequest
     */
    'password'?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof AdminUpdateUserRequest
     */
    'roles'?: Set<string>;
    /**
     * 
     * @type {boolean}
     * @memberof AdminUpdateUserRequest
     */
    'enabled'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AdminUpdateUserRequest
     */
    'accountNonExpired'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AdminUpdateUserRequest
     */
    'accountNonLocked'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AdminUpdateUserRequest
     */
    'credentialsNonExpired'?: boolean;
}