// react-frontend/src/api/models/admin-create-user-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AdminCreateUserRequest
 */
export interface AdminCreateUserRequest {
    /**
     * 
     * @type {string}
     * @memberof AdminCreateUserRequest
     */
    'firstName': string;
    /**
     * 
     * @type {string}
     * @memberof AdminCreateUserRequest
     */
    'lastName': string;
    /**
     * 
     * @type {string}
     * @memberof AdminCreateUserRequest
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof AdminCreateUserRequest
     */
    'password': string;
    /**
     * 
     * @type {Set<string>}
     * @memberof AdminCreateUserRequest
     */
    'roles'?: Set<string>;
    /**
     * 
     * @type {boolean}
     * @memberof AdminCreateUserRequest
     */
    'enabled'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AdminCreateUserRequest
     */
    'accountNonExpired'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AdminCreateUserRequest
     */
    'accountNonLocked'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AdminCreateUserRequest
     */
    'credentialsNonExpired'?: boolean;
}